ARNet_material_classification/requirements.txt
librosa==0.10.0.post2
matplotlib==3.7.1
munch==4.0.0
natsort==8.3.1
numpy==1.24.2
open3d==0.13.0
pandas
pytorch_lightning==2.0.1
PyYAML==6.0.1
scikit_learn
scipy
seaborn==0.12.2
torch
torchmetrics==0.11.4
torchvision
gdown
tensorboard
wandb
openpyxl